# docs/verification/pzs-005-run-1788271963852-3330520/generated-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Themis API",
    "version": "0.1.0",
    "x-pzs005-run-id": "RUN-1788271963852-3330520"
  },
  "paths": {
...
    "/auth/session": {
      "get": {
        "responses": {
          "200": {
            "description": "Current anonymous, restricted, or full session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionEnvelope"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Access denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upstream or mediated boundary returned an invalid response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service temporarily unavailable."
          }
        }
      }
    },
...
    "/test/mailbox/latest": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "$ref": "#/components/schemas/AuthEmail"
            }
          },
          {
            "in": "query",
            "name": "purpose",
            "schema": {
              "type": "string",
              "const": "bootstrap_recovery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest mailbox message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxMessage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Access denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upstream or mediated boundary returned an invalid response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service temporarily unavailable."
          }
        }
      }
    },
    "/test/mailbox": {
      "delete": {
        "responses": {
          "204": {
            "description": "Mailbox cleared."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Access denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Upstream or mediated boundary returned an invalid response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service temporarily unavailable."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthEmail": {
        "type": "string",
        "maxLength": 254,
        "format": "email",
        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
        "description": "Account email address.",
        "example": "engineer@themis.dev"
      },
...
      "RestrictedAuthSession": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "const": "restricted"
          },
          "expiresAt": {
            "type": "string"
          },
          "verifiedEmail": {
            "$ref": "#/components/schemas/AuthEmail"
          }
        },
        "required": ["kind", "expiresAt", "verifiedEmail"],
        "additionalProperties": false
      },
      "AuthSessionEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthSessionResponse"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["data", "message"],
        "additionalProperties": false
      },
      "AuthSessionResponse": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "anonymous"
              }
            },
            "required": ["kind"],
            "additionalProperties": false
          },
          {
            "$ref": "#/components/schemas/RestrictedAuthSession"
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "full"
              },
              "user": {
                "$ref": "#/components/schemas/AuthUser"
              }
            },
            "required": ["kind", "user"],
            "additionalProperties": false
          }
        ],
        "type": "object"
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "Active account identifier.",
            "example": "account-123"
          },
          "email": {
            "$ref": "#/components/schemas/AuthEmail"
          },
          "emailVerifiedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "type": "string",
            "description": "User identifier.",
            "example": "user-123"
          },
          "role": {
            "type": "string",
            "description": "Active account role.",
            "example": "owner"
          },
          "authenticationMethod": {
            "type": "string",
            "const": "passkey"
          },
          "credentialId": {
            "type": "string"
          }
        },
        "required": ["accountId", "email", "emailVerifiedAt", "id", "role"],
        "additionalProperties": false
      },
...
      "MailboxMessage": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string"
          },
          "email": {
            "$ref": "#/components/schemas/AuthEmail"
          },
          "expiresAt": {
            "type": "string"
          },
          "flowId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "pin": {
            "type": "string"
          },
          "purpose": {
            "type": "string",
            "const": "bootstrap_recovery"
          },
          "sentAt": {
            "type": "string"
          }
        },
        "required": ["challengeId", "email", "expiresAt", "flowId", "pin", "purpose", "sentAt"],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      }
    }
  }
}